# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PROSPECTIVE Web API",
    "description": "REST + WebSocket API for cerebral aneurysm segmentation, detection, morphometry, perforator risk assessment and stent planning.\n\nAll linear dimensions are in **millimetres (mm)**. All HU values follow the standard Hounsfield scale.",
    "contact": {
      "name": "SkullApp — Laboratorio de Imagen Médica",
      "email": "ingprospective@skullapp.tech"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/print-prep/beds": {
      "get": {
        "tags": [
          "print-prep"
        ],
        "summary": "List 3D-printer bed presets",
        "operationId": "list_beds_api_print_prep_beds_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrintBed"
                  },
                  "type": "array",
                  "title": "Response List Beds Api Print Prep Beds Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/print-prep/{session_id}": {
      "post": {
        "tags": [
          "print-prep"
        ],
        "summary": "Prepare the vessel mesh for 3D printing",
        "description": "Fill holes → smooth → optional subdivision → scale to a target size, then report physical dimensions, solid volume, watertightness (open-edge count) and whether it fits the requested printer bed. Writes a print-ready STL. Requires that segmentation produced `vessel_tree.vtp`.",
        "operationId": "print_prep_api_print_prep__session_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrintPrepRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintPrepResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/preprocess/{session_id}": {
      "post": {
        "tags": [
          "preprocess"
        ],
        "summary": "Preprocess the session volume (clip / isotropic resample / smooth)",
        "description": "Applies optional HU clipping, isotropic resampling and Gaussian smoothing to the cached volume and rewrites the MPR cache. The segmentation and any downstream analysis must be re-run afterwards. Requires an uploaded volume.",
        "operationId": "preprocess_api_preprocess__session_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreprocessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreprocessResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "preprocess"
        ],
        "summary": "Whether this session's volume has been preprocessed",
        "description": "Lets the panel offer «Revertir» after a session is resumed, when the browser has no memory of the operations that were applied.",
        "operationId": "preprocess_status_api_preprocess__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreprocessStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "preprocess"
        ],
        "summary": "Revert the volume to the original DICOM",
        "description": "Discards the preprocessed volume and rebuilds it from the DICOM still stored in the session — a complete undo of HU clipping, resampling and smoothing, with no re-upload. The segmentation and everything downstream must be re-run, since they were derived from the altered volume.",
        "operationId": "revert_preprocess_api_preprocess__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreprocessResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreprocessRequest": {
        "properties": {
          "clip_hu": {
            "type": "boolean",
            "title": "Clip Hu",
            "description": "Clamp HU outliers to [-1000, 3000]. Ignored on modalities that are not in Hounsfield units (XA/3DRA, MRA), where the same clamp would flatten the contrast column instead of removing outliers.",
            "default": true
          },
          "resample_isotropic": {
            "type": "boolean",
            "title": "Resample Isotropic",
            "description": "Resample to cubic voxels",
            "default": false
          },
          "target_spacing_mm": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.2,
            "title": "Target Spacing Mm",
            "description": "Isotropic voxel size (mm)",
            "default": 0.5
          },
          "smooth": {
            "type": "boolean",
            "title": "Smooth",
            "description": "Mild Gaussian denoising",
            "default": false
          },
          "smooth_sigma": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.1,
            "title": "Smooth Sigma",
            "description": "Gaussian sigma",
            "default": 0.5
          }
        },
        "type": "object",
        "title": "PreprocessRequest"
      },
      "PreprocessResult": {
        "properties": {
          "shape_before": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Shape Before"
          },
          "shape_after": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Shape After"
          },
          "spacing_before": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Spacing Before"
          },
          "spacing_after": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Spacing After"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "description": "What changed; downstream segmentation must be re-run"
          }
        },
        "type": "object",
        "required": [
          "shape_before",
          "shape_after",
          "spacing_before",
          "spacing_after",
          "note"
        ],
        "title": "PreprocessResult"
      },
      "PreprocessStatus": {
        "properties": {
          "applied": {
            "type": "boolean",
            "title": "Applied",
            "description": "True when the cached volume was rewritten"
          },
          "ops": {
            "type": "string",
            "title": "Ops",
            "description": "Human-readable list of the operations applied",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "applied"
        ],
        "title": "PreprocessStatus",
        "description": "Whether this session's volume has been preprocessed, and with what.\n\nA resumed session has no client-side memory of it, so «Revertir» would look\nunavailable on a volume that had in fact been altered."
      },
...
      "PrintBed": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "x_mm": {
            "type": "number",
            "title": "X Mm"
          },
          "y_mm": {
            "type": "number",
            "title": "Y Mm"
          },
          "z_mm": {
            "type": "number",
            "title": "Z Mm"
          }
        },
        "type": "object",
        "required": [
          "name",
          "x_mm",
          "y_mm",
          "z_mm"
        ],
        "title": "PrintBed"
      },
      "PrintPrepRequest": {
        "properties": {
          "target_size_mm": {
            "type": "number",
            "maximum": 400.0,
            "minimum": 0.0,
            "title": "Target Size Mm",
            "description": "Max dimension after uniform scaling (mm). 0 = keep original size.",
            "default": 80.0
          },
          "smooth_iterations": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "Smooth Iterations",
            "description": "Laplacian smoothing iterations",
            "default": 20
          },
          "smooth_relaxation": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Smooth Relaxation",
            "description": "Smoothing relaxation factor",
            "default": 0.1
          },
          "fill_holes": {
            "type": "boolean",
            "title": "Fill Holes",
            "description": "Fill holes before smoothing",
            "default": true
          },
          "hole_size": {
            "type": "number",
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Hole Size",
            "description": "Max hole perimeter to fill (mm)",
            "default": 5.0
          },
          "subdivide": {
            "type": "boolean",
            "title": "Subdivide",
            "description": "One level of linear subdivision (denser mesh)",
            "default": false
          },
          "bed_x_mm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Bed X Mm",
            "description": "Print bed X (mm); 0 = unlimited",
            "default": 0.0
          },
          "bed_y_mm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Bed Y Mm",
            "description": "Print bed Y (mm); 0 = unlimited",
            "default": 0.0
          },
          "bed_z_mm": {
            "type": "number",
            "minimum": 0.0,
            "title": "Bed Z Mm",
            "description": "Print bed Z (mm); 0 = unlimited",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "PrintPrepRequest",
        "description": "Prepare the working vessel mesh for 3D printing."
      },
      "PrintPrepResult": {
        "properties": {
          "stl_url": {
            "type": "string",
            "title": "Stl Url",
            "description": "URL of the print-ready STL"
          },
          "scale_factor": {
            "type": "number",
            "title": "Scale Factor"
          },
          "dimensions_mm": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Dimensions Mm",
            "description": "[dx, dy, dz] after scaling"
          },
          "volume_cm3": {
            "type": "number",
            "title": "Volume Cm3"
          },
          "surface_area_cm2": {
            "type": "number",
            "title": "Surface Area Cm2"
          },
          "is_watertight": {
            "type": "boolean",
            "title": "Is Watertight"
          },
          "open_edge_count": {
            "type": "integer",
            "title": "Open Edge Count"
          },
          "fits_in_bed": {
            "type": "boolean",
            "title": "Fits In Bed",
            "description": "Whether the model fits the requested bed"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "stl_url",
          "scale_factor",
          "dimensions_mm",
          "volume_cm3",
          "surface_area_cm2",
          "is_watertight",
          "open_edge_count",
          "fits_in_bed"
        ],
        "title": "PrintPrepResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}